move distance sql to a separate script
